CSP: Should only honor CSP policy delivered in meta tag that is a descendent of <head> https://bugs.webkit.org/show_bug.cgi?id=59858 <rdar://problem/25603538>
Reviewed by Brent Fulgham.

Source/WebCore:

Ignore the Content Security Policy meta tag if it is not a descendent of <head> as per
section HTML meta Element of the Content Security Policy Level 2 spec., <https://w3c.github.io/webappsec-csp/2/>
(Editor's Draft, 29 August 2015).

Tests: http/tests/security/contentSecurityPolicy/meta-tag-ignored-if-not-in-head.html
       http/tests/security/contentSecurityPolicy/meta-tag-ignored-if-not-in-head2.html
       http/tests/security/contentSecurityPolicy/report-only-meta-tag-ignored-if-not-in-head.html
       http/tests/security/contentSecurityPolicy/report-only-meta-tag-ignored-if-not-in-head2.html

* dom/Document.cpp:
(WebCore::Document::processHttpEquiv): Modified to take a boolean argument whether the http-equiv
meta tag is a descendent of <head> and to parse the value of a Content Security Policy http-equiv
only if the http-equiv meta tag is a descendent of <head>.
* dom/Document.h: Add parameter isInDocument to processHttpEquiv(). Remove javadoc-style parameters
from processHttpEquiv() comment as we do not document parameters for non-API functions using such style.
Also write the comment for processHttpEquiv() using C++ style comments instead of a C-style comment.
* html/HTMLMetaElement.cpp:
(WebCore::HTMLMetaElement::process): Pass whether this element is a descendent of <head>. Additionally
update stale comment and move it closer to the code it refers to.

LayoutTests:

Add tests to ensure that we ignore the meta tags for Content-Security-Policy, Content-Security-Policy-Report-Only,
X-WebKit-CSP, and X-WebKit-CSP-Report-Only if it is not a descendent of <head>.

* http/tests/security/contentSecurityPolicy/meta-tag-ignored-if-not-in-head-expected.txt: Added.
* http/tests/security/contentSecurityPolicy/meta-tag-ignored-if-not-in-head.html: Added.
* http/tests/security/contentSecurityPolicy/meta-tag-ignored-if-not-in-head2-expected.txt: Added.
* http/tests/security/contentSecurityPolicy/meta-tag-ignored-if-not-in-head2.html: Added.
* http/tests/security/contentSecurityPolicy/report-only-meta-tag-ignored-if-not-in-head-expected.txt: Added.
* http/tests/security/contentSecurityPolicy/report-only-meta-tag-ignored-if-not-in-head.html: Added.
* http/tests/security/contentSecurityPolicy/report-only-meta-tag-ignored-if-not-in-head2-expected.txt: Added.
* http/tests/security/contentSecurityPolicy/report-only-meta-tag-ignored-if-not-in-head2.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@199163 268f45cc-cd09-0410-ab3c-d52691b4dbfc