Web Inspector: Uncaught Exception: No resource with given URL found https://bugs.webkit.org/show_bug.cgi?id=156259 <rdar://problem/25564749>
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2016-04-07
Reviewed by Timothy Hatcher.

* UserInterface/Views/SourceCodeTextEditor.js:
(WebInspector.SourceCodeTextEditor):
SourceCode.prototype.requestContent will reject if it cannot load
content for the given resource. In that case, we already have an
earlier catch handler that displays an error message in the
ContentView, so we shouldn't show an Uncaught Exception page.
Really, we should not reject the original promise here, and
instead resolve it with an object describing the error, but
short term we should remove the uncaught exception handler for
this case.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@199178 268f45cc-cd09-0410-ab3c-d52691b4dbfc