Rename ArrayMode::supportsLength to supportsSelfLength https://bugs.webkit.org/show_bug.cgi?id=156374
Reviewed by Filip Pizlo.

The name supportsLength is confusing because TypedArray have a
length function however it is on the prototype and not on the
instance. supportsSelfLength makes more sense since we use the
function during fixup to tell if we can intrinsic the length
property lookup on self accesses.

* dfg/DFGArrayMode.h:
(JSC::DFG::ArrayMode::supportsSelfLength):
(JSC::DFG::ArrayMode::supportsLength): Deleted.
* dfg/DFGFixupPhase.cpp:
(JSC::DFG::FixupPhase::attemptToMakeGetArrayLength):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@199203 268f45cc-cd09-0410-ab3c-d52691b4dbfc