[iOS] Play button on video is too dark https://bugs.webkit.org/show_bug.cgi?id=156383 <rdar://problem/23540816>
Reviewed by Simon Fraser.

.:

Add a manual test for iOS that shows the expected appearance
of a video element. Unfortunately, due to the way we take
snapshots on iOS within our test runner, we don't get the
platform blurring effect, which means an automated test
won't work.

* ManualTests/ios/start-playback-button-appearance-expected.html: Added.
* ManualTests/ios/start-playback-button-appearance.html: Added.

Source/WebKit2:

Elements that are backed by a layer with either LightBackdropAppearance
or DarkBackdropAppearance are actually a combination of a
few layers (inside a special view). If we apply a mask to one
of those layers, it needs to be attached to the correct
child layer.

* Shared/mac/RemoteLayerTreePropertyApplier.mm:
(WebKit::RemoteLayerTreePropertyApplier::applyProperties): If we have
one of the special appearance flags, apply the mask layer to
a particular child, rather than the layer itself.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@199212 268f45cc-cd09-0410-ab3c-d52691b4dbfc