JSContext Inspector: Fix asserts and uncaught exception showing Timeline Tab https://bugs.webkit.org/show_bug.cgi?id=156411
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2016-04-08
Reviewed by Timothy Hatcher.

* UserInterface/Views/OverviewTimelineView.js:
(WebInspector.OverviewTimelineView):
(WebInspector.OverviewTimelineView.prototype.closed):
Gracefully handle if we do not have a Network Timeline.

* UserInterface/Views/TimelineTabContentView.js:
(WebInspector.TimelineTabContentView.prototype._changeViewMode):
This function is always called by the constructor, so the assert
is not useful since it can be called when FPS is not supported.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@199241 268f45cc-cd09-0410-ab3c-d52691b4dbfc