[CMake] Make FOLDER property INHERITED https://bugs.webkit.org/show_bug.cgi?id=156460
Patch by Fujii Hironori <Hironori.Fujii@jp.sony.com> on 2016-04-11
Reviewed by Brent Fulgham.

.:

Some CMake targets are not setting the FOLDER property. This causes the
generated projects to be displayed in the top-level folder of the solution.

Making the FOLDER property INHERITED ensures that all the targets
are placed in their proper directories.

* Source/cmake/OptionsCommon.cmake:
Define FOLDER property as a inherited property.
* Source/cmake/WebKitMacros.cmake:
Do not set FOLDER target property.

Source/bmalloc:

* CMakeLists.txt:
Set FOLDER property as a directory property not a target property

Source/JavaScriptCore:

* CMakeLists.txt:
* shell/CMakeLists.txt:
* shell/PlatformWin.cmake:
Set FOLDER property as a directory property not a target property

Source/ThirdParty/ANGLE:

* CMakeLists.txt:
Set FOLDER property as a directory property not a target property

Source/WebCore:

* CMakeLists.txt:
Set FOLDER property as a directory property not a target property

Source/WebKit:

* CMakeLists.txt:
* PlatformWin.cmake:
Set FOLDER property as a directory property not a target property

Source/WebKit2:

* CMakeLists.txt:
Set FOLDER property as a directory property not a target property

Source/WTF:

* CMakeLists.txt:
Set FOLDER directory property

Tools:

* CMakeLists.txt:
* DumpRenderTree/CMakeLists.txt:
* DumpRenderTree/PlatformWin.cmake:
* ImageDiff/CMakeLists.txt:
* MiniBrowser/efl/CMakeLists.txt:
* MiniBrowser/gtk/CMakeLists.txt:
* MiniBrowser/win/CMakeLists.txt:
Set FOLDER property as a directory property not a target property

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@199289 268f45cc-cd09-0410-ab3c-d52691b4dbfc