ShadowChicken::visitChildren() should not visit tailMarkers and throwMarkers. https://bugs.webkit.org/show_bug.cgi?id=156532
Reviewed by Saam Barati and Filip Pizlo.

ShadowChicken can store tailMarkers and throwMarkers in its log, specifically in
the callee field of a log packet.  However, ShadowChicken::visitChildren()
unconditionally visits the callee field of each packet as if they are real
objects.  If visitChildren() encounters one of these markers in the log, we get a
crash.

This crash was observed in the v8-v6/v8-regexp.js stress test running with shadow
chicken when r199393 landed.  r199393 introduced tail calls to a RegExp split
fast path, and the v8-regexp.js test exercised this fast path a lot.  Throw in
some timely GCs, and we get a crash party.

The fix is to have ShadowChicken::visitChildren() filter out the tailMarker and
throwMarker.

Alternatively, if perf is an issue, we can allocate 2 dedicated objects for
these markers so that ShadowChicken can continue to visit them.  For now, I'm
going with the filter.

* interpreter/ShadowChicken.cpp:
(JSC::ShadowChicken::visitChildren):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@199496 268f45cc-cd09-0410-ab3c-d52691b4dbfc