Unindent an unnecessary block in stringProtoFuncSplitFast(). https://bugs.webkit.org/show_bug.cgi?id=156802
Reviewed by Filip Pizlo.

In webkit.org/b/156013, I refactored stringProtoFuncSplit into
stringProtoFuncSplitFast.  In that patch, I left an unnecessary block of code in
its original block (with FIXMEs) to keep the diff for that patch minimal.  Now
that the patch for webkit.org/b/156013 has landed, I will unindent that block and
remove the FIXMEs.

* runtime/StringPrototype.cpp:
(JSC::stringProtoFuncSplitFast):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@199780 268f45cc-cd09-0410-ab3c-d52691b4dbfc