[css-grid] Fix bug with positioned items in vertical writing mode https://bugs.webkit.org/show_bug.cgi?id=156870
Reviewed by Darin Adler.

Source/WebCore:

In RenderGrid::offsetAndBreadthForPositionedChild() we were using
directly borderLeft(), which is wrong in vertical writing modes.

To fix it we just need to use borderLogicalLeft() which is aware of
the current writing mode.

Test: fast/css-grid-layout/grid-positioned-children-writing-modes.html

* rendering/RenderGrid.cpp:
(WebCore::RenderGrid::offsetAndBreadthForPositionedChild):

LayoutTests:

Add new test to check positioned items in different writing modes
and direction combinations.

* fast/css-grid-layout/grid-positioned-children-writing-modes-expected.html: Added.
* fast/css-grid-layout/grid-positioned-children-writing-modes.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@199874 268f45cc-cd09-0410-ab3c-d52691b4dbfc