Clean up Font::removeFromSystemFallbackCache() https://bugs.webkit.org/show_bug.cgi?id=157093
Reviewed by Darin Adler.

Fonts have a bit which represents whether or not the system fallback cache references them.
This bit gets set to 1 when they are inserted into the system fallback cache, but was not
being reset back to 0 when it was removed.

This is benign, since the old codepath has no effect. However, we can speed up the function
by properly updating this bit.

No new tests because there is no behavior change.

* platform/graphics/Font.cpp:
(WebCore::Font::removeFromSystemFallbackCache):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@200150 268f45cc-cd09-0410-ab3c-d52691b4dbfc