[JSC] In DFG, an OSR Exit on SetLocal can trash its child node https://bugs.webkit.org/show_bug.cgi?id=157358 rdar://problem/25339647
Patch by Benjamin Poulain <bpoulain@apple.com> on 2016-05-05
Reviewed by Filip Pizlo.

When we OSR Exit on SetLocal, the child is never restored if its representation
was changed since the MovHint.

For example, say we have:
    @1 = SomethingProducingDouble()
    @2 = MovHint(@1)
    @3 = ValueRep(@1)
    @4 = SetLocal(@3, FlushedInt32)

When we lower SetLocal(), we start by speculating that @3 is an Int32.
Now this can fail if @1 was really a double.
When that happens, we go over the VariableEventStream to find where values
are, and @1 died at @3. Since the speculation failure happens before
the SetLocal event, we don't do anything with @3.

In this patch, I extend the PhantomInsertion phase to keep the MovHint
alive past the SetLocal.

* dfg/DFGPhantomInsertionPhase.cpp:
* tests/stress/multiply-typed-double-and-object.js: Added.
(otherObject.valueOf):
(targetDFG.multiply):
(targetFTL.multiply):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@200498 268f45cc-cd09-0410-ab3c-d52691b4dbfc