Handle _schemeUpgraded delegate callbacks in NSURLSessionDataDelegate https://bugs.webkit.org/show_bug.cgi?id=157354 rdar://problem/25842107
Reviewed by Darin Adler.

Source/WebCore:

No new tests.  This cannot be tested with a self-signed certificate.
This needs to be tested once we use real ssl certificates for testing.

* platform/network/mac/WebCoreResourceHandleAsDelegate.mm:
(-[WebCoreResourceHandleAsDelegate connection:willSendRequest:redirectResponse:]):
* platform/network/mac/WebCoreResourceHandleAsOperationQueueDelegate.mm:
(-[WebCoreResourceHandleAsOperationQueueDelegate connection:willSendRequest:redirectResponse:]):
* platform/network/mac/WebCoreURLResponse.h:
* platform/network/mac/WebCoreURLResponse.mm:
(WebCore::synthesizeRedirectResponseIfNecessary):
Take the NSURLRequest instead of the NSURLConnection as a parameter so we can share this code with the NSURLSession loader,
which has an NSURLSessionDataTask.  Both have the currentRequest, which is all we need in this function anyway.

Source/WebKit2:

* NetworkProcess/cocoa/NetworkSessionCocoa.mm:
(-[WKNetworkSessionDelegate URLSession:task:_schemeUpgraded:completionHandler:]):
When we make a request to a http url and CFNetwork determines that that url would have redirected to an HSTS site
and it is going to change the request to an https request, this delegate callback is called.  We need to call the
redirection code to have the same behavior as the NSURLConnection-based loader.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@200655 268f45cc-cd09-0410-ab3c-d52691b4dbfc