Uncaught Exception: TypeError: null is not an object (evaluating 'event.data.pathComponent.domTreeElement') https://bugs.webkit.org/show_bug.cgi?id=157759 <rdar://problem/26309427>
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2016-05-16
Reviewed by Timothy Hatcher.

* UserInterface/Views/DOMTreeContentView.js:
(WebInspector.DOMTreeContentView.prototype._pathComponentSelected):
Handle possible null path components.

* UserInterface/Views/HierarchicalPathComponent.js:
(WebInspector.HierarchicalPathComponent.prototype.get selectedPathComponent):
When there is a list of just one that is triggered, the Clicked
event should produce that one, to be consistent with clicking
on a list of multiple elements and selecting the same value.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@200995 268f45cc-cd09-0410-ab3c-d52691b4dbfc