Mail crashes under TiledCoreAnimationDrawingAreaProxy::createFenceForGeometryUpdate() when the Web process crashes https://bugs.webkit.org/show_bug.cgi?id=157927 <rdar://problem/24793374>
Reviewed by Dean Jackson.

* UIProcess/mac/TiledCoreAnimationDrawingAreaProxy.mm:
(WebKit::TiledCoreAnimationDrawingAreaProxy::createFence):
With just the right combination of layout mechanisms and Web Content process
crash timing, it is possible to get a frame change underneath WebPageProxy::processDidCrash,
before the DrawingArea has been nulled out but after the Connection has.
This can lead to a null dereference. I had a lot of trouble writing a API
test for this, though I'm certain it's possible.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@201202 268f45cc-cd09-0410-ab3c-d52691b4dbfc