Web Inspector: Visual Style connectors are not always centered properly https://bugs.webkit.org/show_bug.cgi?id=157932
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2016-05-19
Reviewed by Timothy Hatcher.

* UserInterface/Views/VisualStylePropertyEditorLink.css:
(.visual-style-property-editor-link):
Remove the 1px margin on @2x that would cause jitter at different
sidebar widths. It is however needed on @1x.

(.visual-style-property-editor-link:not(.link-all)):
Make the height 20 so these both move closer to the center of the
<select> elements they are next to. They are not perfectly centered
as that would require a half pixel.

(.visual-style-property-editor-link.link-vertical):
Move the top line up a pixel so it is "up 0.5" from center and
the bottom one is already "down 0.5" from the center.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@201209 268f45cc-cd09-0410-ab3c-d52691b4dbfc